LibTLS: Move out Certificate to its own header file
